+++ Niverobot/Niverobot.Domain/EfModels/NiveroBotContext.cs
@@ -1,0 +1,18 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+
+namespace Niverobot.Domain.EfModels
+{
+    public class NiveroBotContext : DbContext
+    {
+        public NiveroBotContext(DbContextOptions<NiveroBotContext> options)
+            : base(options)
+        {
+        }
+        
+        public DbSet<Reminder> Reminders { get; set; }
+
+
+
+    }
+}