+++ Niverobot/Niverobot.Interfaces/IDadJokeService.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Niverobot.Interfaces
+{
+    public interface IDadJokeService
+    {
+        Task<string> GetDadJokeAsync();
+    }
+}