+++ Niverobot/Niverobot.Interfaces/IGRPCService.cs
@@ -1,0 +1,9 @@
+﻿using Dateparser;
+
+namespace Niverobot.Interfaces
+{
+    public interface IGRPCService
+    {
+        ParseDateReply ParseDateTimeFromNl(string date);
+    }
+}