+++ Niverobot/Niverobot.Interfaces/ITimezoneService.cs
@@ -1,0 +1,7 @@
+﻿namespace Niverobot.Interfaces
+{
+    public interface ITimezoneService
+    {
+        int GetUtcOffsetInSeconds(double lat, double lon);
+    }
+}