+++ Niverobot/Niverobot.Services/TelegramUpdateService.cs
@@ -1,21 +1,17 @@
-using Niverobot.WebApi.Interfaces;
-using Serilog;
-using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
+using Niverobot.Interfaces;
+using Serilog;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 
-namespace Niverobot.WebApi.Services
+namespace Niverobot.Services
 {
     public class TelegramUpdateService : ITelegramUpdateService
     {
-        private readonly ITelegramBotService _botService;
         private readonly IMessageService _messageService;
 
-        public TelegramUpdateService(ITelegramBotService botService, IMessageService messageService)
+        public TelegramUpdateService(IMessageService messageService)
         {
-            _botService = botService;
             _messageService = messageService;
         }